pyproject.toml
[tool.poetry]
name = "august-intern"
version = "0.1.0"
description = ""
authors = ["Juunsik <Juunsik29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
pytest-django = "^4.8.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
